player interactor interface was added
