+++ app/src/main/java/bes/max/playlistmaker/presentation/player/PlayerViewModelFactory.kt
@@ -4,14 +4,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import bes.max.playlistmaker.di.Creator
 import bes.max.playlistmaker.domain.models.Track
-import bes.max.playlistmaker.domain.player.PlayerInteractor
+import bes.max.playlistmaker.domain.player.PlayerInteractorImpl
 
 class PlayerViewModelFactory(private val track: Track): ViewModelProvider.Factory {
     private val creator = Creator(null)
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(PlayerViewModel::class.java)) {
-            return PlayerViewModel(track, PlayerInteractor(creator.getNewPlayerImpl())) as T
+            return PlayerViewModel(track, PlayerInteractorImpl(creator.getNewPlayerImpl())) as T
         }
         throw IllegalArgumentException("Unknown ViewModel")
     }
